problem 1 marked

+++ cpp/ps0001/ps00001.cpp
@@ -3,12 +3,17 @@ using namespace std;
 
 int main()
 {
-  int num1;
-  int num2;
-  int num3;
-  cin >>num1;
+  // prompting user for 3 numbers to add 1/2
+  unsigned short num1;
+  unsigned short num2;
+  unsigned short num3;
+  cin >> num1;
   cin >> num2;
   cin >> num3;
-  cout<<"The Solution Is\n"<<num1+num2+num3;
+
+  // outputting sum of the 3 numbers 3/3
+  cout << "The Solution Is\n" << num1 + num2 + num3;
+  
   return 0;
 }
+// great work 😎 8/10 ⭐
